1. renamed go-redis-v8 to goredis8
