+++ drivers/goredis8/factory.go
@@ -1,4 +1,4 @@
-package go_redis_v8
+package goredis8
 
 import (
 	"context"
