+++ drivers/goredis8/readonly_func_without_tx.go
@@ -1,4 +1,4 @@
-package go_redis_v8
+package goredis8
 
 import (
 	"context"
